requirements.txt
attrs==21.2.0
Automat==20.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2021.5.30
cffi==1.14.6
cleo==0.8.1
clikit==0.6.2
constantly==15.1.0
crashtest==0.3.1
cryptography==35.0.0
cssselect==1.1.0
dataclasses==0.8
decorator==5.1.0
distlib==0.3.3
filelock==3.3.1
greenlet==1.1.2
h2==3.2.0
hpack==3.0.0
html5lib==1.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==1.7.0
importlib-resources==5.2.2
incremental==21.3.0
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
itemadapter==0.4.0
itemloaders==1.0.4
jedi==0.17.2
jeepney==0.7.1
jmespath==0.10.0
keyring==21.8.0
lockfile==0.12.2
lxml==4.6.3
msgpack==1.0.2
numpy==1.19.5
packaging==20.9
pandas==1.1.5
parsel==1.6.0
parso==0.7.1
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.7.1
platformdirs==2.4.0
poetry==1.1.11
poetry-core==1.0.7
priority==1.3.0
prompt-toolkit==3.0.20
Protego==0.1.16
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.10.0
pylev==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
queuelib==1.6.2
requests-toolbelt==0.9.1
Scrapy==2.5.1
SecretStorage==3.3.1
service-identity==21.1.0
shellingham==1.4.0
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
tomlkit==0.7.2
traitlets==4.3.3
Twisted==21.7.0
typing-extensions==3.10.0.2
virtualenv==20.8.1
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.6.0
zope.interface==5.4.0
